Make HTTPSConnection.close() actually close underlying socket.
Originally HTTPSConnection.close() was intentinoall made into NOOP to
avoid line 545 of httplib.py,
https://hg.python.org/cpython/file/2.3/Lib/httplib.py#l545 where in
HTTPConnection.connect() self.sock can be closed in the exception
handler, yet the connection would still remain open in the response
object.

In Python 2.6 (and later) that exception handling is moved to
the socket library (where it belongs), so I believe the above
kludge is not required anymore.

Fixes #203